Total rework of data types. Now datatypes have its own domain
Generated code from xml tend to be a bad practice for an api delivered
as a dependency for others to use. How code is generated can be very
unstable with regards to convensions for Snake_Case, Capitalize etc.
It also offers no support for serialisation to xml and dependents need to
have data added in a non typed fashion (eg. string). I also lacs
support for helper-functions and correct constructor logic.

Therefore I have now created a new domain that shall be used. I have not
implemented DataTypes for types typically only used by Digipost (Boligmappa etc.)
This can be done on demand by request.

I removed the module for generating code and now use xscgen in stead.

I have upgraded the project to net6.0, but datatypes are still
released on netstandard2.0.